refactor(onboarding): restructure and rename components
