+++ renderer/components/Onboarding/Steps/components/BackupTypeItem.js
