+++ renderer/components/Onboarding/Steps/components/ConnectionTypeItem.js
