+++ renderer/components/Onboarding/Steps/index.js
@@ -12,5 +12,4 @@ export SeedConfirm from './SeedConfirm'
 export SeedView from './SeedView'
 export WalletCreate from './WalletCreate'
 export WalletRecover from './WalletRecover'
-export ConnectionTypeItem from './ConnectionTypeItem'
-export SetupBackup from './SetupBackup'
+export BackupSetup from './BackupSetup'
